refactor(api): Se crea aplicación
Se crea una aplicación específica para la API ordenando mejor el
proyecto
