+++ searcher/views.py
@@ -9,12 +9,9 @@ from django.conf import settings
 from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.contrib import messages
 from .helpers import (data_processor, variant_to_query, ethno_table_maker,
-                      query_kreator, get_variants, results_to_csv,
-                      api_data_processor)
+                      query_kreator, get_variants, results_to_csv)
 from .forms import SearchPostForm
 from elasticsearch import Elasticsearch
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
 
 LOGGER = logging.getLogger(__name__)
 
@@ -125,41 +122,6 @@ def search(request):
     else:
         # Si es metodo GET se redirige a la vista index
         return HttpResponseRedirect('/')
-
-
-@api_view(['GET', 'POST'])
-def search_api(request):
-    """Endpoint para cpnsultas por medio de peticiones POST"""
-    entries_count = 0
-    if request.method == "POST":
-        variants = ""
-        data = request.data
-        lang = data["lang"]
-        if "variants" in data.keys() and data["variant"]:
-            variants = " AND variant:" + variant_to_query(data["variants"])
-        query = query_kreator(f"{lang}:{data['query']}{variants}")
-        r = es.search(index=settings.INDEX, body=query, scroll="1m")
-        data_response = r["hits"]
-        scroll_id = r["_scroll_id"]
-        total_entries = data_response["total"]["value"]
-        entries_count = len(data_response["hits"])
-        while entries_count != total_entries:
-            sub_response = es.scroll(scroll_id=scroll_id, scroll="1m")
-            data_response["hits"] += sub_response["hits"]["hits"]
-            entries_count += len(sub_response["hits"]["hits"])
-            scroll_id = sub_response["_scroll_id"]
-        preprocess_data = api_data_processor(data_response)
-        return Response({"query": data["query"],
-                         "results": preprocess_data,
-                         "elasticsearch-index": settings.INDEX,
-                         "total_results": total_entries})
-    current_variants = get_variants()
-    del current_variants["status"]
-    return Response({"variants": current_variants, "l1": settings.L1,
-                     "l2": settings.L2})
-
-
-# === Datos de Ethnologue ===
 
 
 def ethnologue_data(request, iso_variant):
